Remove unnecessary function for creating UserForExport object
